removed cors, added copy command into /dist, admin.ts needs some work
