+++ frontend/public/index.html
@@ -2,9 +2,9 @@
 <html>
 
 <head>
-    <link rel="stylesheet" href="styles.css" />
+    <link rel="stylesheet" href="./styles.css" />
     <title>Namaz Vakitleri</title>
-    <link rel="icon" href="images/moons/sd3.svg" type="image/x-icon">
+    <link rel="icon" href="./../images/moons/sd3.svg" type="image/x-icon">
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link
         href="https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap"
@@ -12,7 +12,7 @@
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <meta charset="UTF-8" />
     <script src="https://d3js.org/d3.v4.js"></script>
-    <script src="app.js" defer></script>
+    <script src="./../src/app.js" defer></script>
     <meta http-equiv="Permissions-Policy" content="interest-cohort=()">
 </head>
 
@@ -26,19 +26,19 @@
 
         <div class="l-1">
             <div class="l-1-1">
-                <img class="moon1" src="images/moons/sd3.svg">
+                <img class="moon1" src="./../images/moons/sd3.svg">
             </div>
             <div class="l-1-2">
-                <img class="moon2" src="images/moons/sd3.svg">
+                <img class="moon2" src="./../images/moons/sd3.svg">
             </div>
             <div class="l-1-3">
-                <img class="moon3" src="images/moons/sd3.svg">
+                <img class="moon3" src="./../images/moons/sd3.svg">
             </div>
             <div class="l-1-4">
-                <img class="moon4" src="images/moons/sd3.svg">
+                <img class="moon4" src="./../images/moons/sd3.svg">
             </div>
             <div class="l-1-5">
-                <img class="moon5" src="images/moons/sd3.svg">
+                <img class="moon5" src="./../images/moons/sd3.svg">
 
             </div>
         </div>
@@ -124,7 +124,7 @@
                 <style type="text/css">
                     @font-face {
                         font-family: "Hafs";
-                        src: url("../../hafs.otf");
+                        src: url("hafs.otf");
                         font-weight: normal;
                     }
                 </style>
@@ -175,7 +175,7 @@
                 <style type="text/css">
                     @font-face {
                         font-family: "Hafs";
-                        src: url("../../hafs.otf");
+                        src: url("hafs.otf");
                         font-weight: normal;
                     }
                 </style>
@@ -245,7 +245,7 @@
                 <style type="text/css">
                     @font-face {
                         font-family: "Hafs";
-                        src: url("../../hafs.otf");
+                        src: url("hafs.otf");
                         font-weight: normal;
                     }
                 </style>
@@ -297,7 +297,7 @@
                 <style type="text/css">
                     @font-face {
                         font-family: "Hafs";
-                        src: url("../../hafs.otf");
+                        src: url("hafs.otf");
                         font-weight: normal;
                     }
                 </style>
@@ -350,7 +350,7 @@
                 <style type="text/css">
                     @font-face {
                         font-family: "Hafs";
-                        src: url("../../hafs.otf");
+                        src: url("hafs.otf");
                         font-weight: normal;
                     }
                 </style>
@@ -420,7 +420,7 @@
                 <style type="text/css">
                     @font-face {
                         font-family: "Hafs";
-                        src: url("../../hafs.otf");
+                        src: url("hafs.otf");
                         font-weight: normal;
                     }
                 </style>
