+++ frontend/public/versesAndHadiths.json
