+++ package.json
@@ -3,11 +3,10 @@
     "version": "1.0.0",
     "description": "",
     "type": "module",
-    "main": "index.js",
+    "main": "./dist/index.js",
     "scripts": {
-        "test": "echo \"Error: no test specified\" && exit 1",
-        "build": "tsc",
-        "start": "node index.js"
+        "build": "mkdir -p ./dist/frontend && rsync -av --exclude='*.ts' ./frontend/ ./dist/frontend/ && tsc",
+        "start": "node ."
     },
     "author": "",
     "license": "ISC",
